Extract syncer_metrics.go from compactor.go.
